Restructure tractive
