+++ lib/tractive/migrator/converter.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "converter/trac_to_github"
