+++ spec/spec_helper.rb
@@ -3,6 +3,10 @@
 require "tractive"
 require "webmock/rspec"
 
+require_relative "support/helpers/stub_git_api"
+require_relative "support/helpers/common_functions"
+require_relative "support/helpers/ticket_compose"
+
 WebMock.disable_net_connect!(allow_localhost: true)
 CONFIG = YAML.load_file("spec/files/test.config.yaml")
 
@@ -17,6 +21,10 @@ RSpec.configure do |config|
     Tractive::Utilities.setup_logger(verbose: false)
     Tractive::Utilities.setup_db!(CONFIG["trac"]["database"])
   end
+
+  config.include Helpers::StubGitApi
+  config.include Helpers::CommonFunctions
+  config.include Helpers::TicketCompose
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
